fix: update url to be superset.apache.org extract functionallity from glossaryUtils into tooltipUtils
