+++ superset-frontend/packages/superset-ui-core/src/glossary/glossaryUtils.ts
@@ -19,13 +19,6 @@
 
 import { GlossaryMap, GlossaryTerm, type GlossaryTopic } from './glossaryModels';
 import { glossaryDefinition } from './glossary';
-import { t } from '@superset-ui/core';
-
-export const GLOSSARY_BASE_URL = 'http://localhost:3000/docs'; // TODO: Change to use the url for the environment.
-
-// Pattern matches: [GLOSSARY]|topic|title
-// Captures: topic and title for lookup in glossary
-const GLOSSARY_ENCODING_PATTERN = /^\[GLOSSARY\]\|([^|]+)\|([^|]+)$/;
 
 /**
  * The exported glossary object is a runtime structure where each entry is a GlossaryTerm instance, but the key
@@ -57,26 +50,6 @@ const glossary: Glossary = Object.fromEntries(
 const glossaryMap = new GlossaryMap(glossary);
 
 export const getAllGlossaryTopics = (): GlossaryTopic[] => glossaryMap.getAllTopics();
-
-const buildGlossaryUrl = (topic: string, title: string): string =>
-  `${GLOSSARY_BASE_URL}/glossary#${encodeURIComponent(topic + '__' + title)}`;
-
-export const resolveGlossaryString = (glossaryString: string): [string | undefined, string] => {
-  const match = glossaryString.match(GLOSSARY_ENCODING_PATTERN);
-  if (!match) {
-    return [undefined, glossaryString];
-  }
-  const topic = match[1];
-  const title = match[2];
-  
-  // Look up the term from the glossary to get the translated description
-  const glossaryTopic = glossaryMap.getTopic(topic);
-  const term = glossaryTopic?.getTerm(title);
-  const description = term ? term.getShort(t) : glossaryString;
-  
-  const glossaryUrl = buildGlossaryUrl(topic, title);
-  return [glossaryUrl, description];
-};
 
 export const getGlossaryTopic = (topicName: string): GlossaryTopic | undefined =>
   glossaryMap.getTopic(topicName);
